# server/swagger.yaml
openapi: 3.0.3
info:
  title: AVO with Mie - Swagger Page
  description: |-
    This is a server for AVO with Mie website - Website for listening podcast and reading book. 
    
    Some useful links:
    - [The Pet Store repository](https://github.com/ducdatit2002/avowithmie-server)
    - [The source API definition](https://avowithmie-server.onrender.com/api-docs)

  contact:
    email: ducdatit2002@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: https://avowithmie-server.onrender.com/api/
  - url: http://localhost:4000/api/
...
paths:
...
  /podcasts:
    post:
      tags:
        - Podcasts
      summary: Create a new podcast
      description: Admin can create a new podcast entry.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: Token required to authenticate the admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Podcast'
      responses:
        '200':
          description: Podcast created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Podcast'
                  message:
                    type: string
                    example: 'Podcast created successfully'
      security:
        - adminApiKey: []
    get:
      tags:
        - Podcasts
      summary: Get all podcasts
      description: Retrieve all podcasts.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: Token required to access the list of podcasts.
      responses:
        '200':
          description: List of podcasts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Podcast'
      security:
        - apiKeyAuth: []
  /podcasts/{id}:
    put:
      tags:
        - Podcasts
      summary: Update a podcast
      description: Admin can update an existing podcast entry.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: Podcast ID
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: Token required for authentication as an admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Podcast'
      responses:
        '200':
          description: Podcast updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Podcast'
                  message:
                    type: string
                    example: 'Update podcast successfully'
      security:
        - adminApiKey: []

    delete:
      tags:
        - Podcasts
      summary: Delete a podcast
      description: Admin can delete a podcast entry.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: Podcast ID
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: Token required for administrative privileges.
      responses:
        '200':
          description: Podcast deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Podcast deleted successfully'
      security:
        - adminApiKey: []
  /podcasts/like/{id}:
    put:
      tags:
        - Podcasts
      summary: Like or unlike a podcast
      description: User can like or unlike a podcast.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: Podcast ID
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: Token required for user authentication.
      responses:
        '200':
          description: Like status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Added to your liked podcasts'  # This message can change based on the operation performed
      security:
        - auth: []
  /podcasts/like:
    get:
      tags:
        - Podcasts
      summary: Get all liked podcasts
      description: Retrieve all podcasts liked by the user.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: Token required for user authentication.
      responses:
        '200':
          description: List of liked podcasts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Podcast'
      security:
        - auth: []
  /playlists:
    post:
      tags:
        - Playlist
      summary: Create playlist
      description: Create a new playlist.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to authenticate the request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Playlist'
      responses:
        '201':
          description: Playlist created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Playlist'
      security:
        - apiKeyAuth: []

    get:
      tags:
        - Playlist
      summary: Get all playlists
      description: Retrieve all playlists.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to access the list of playlists.
      responses:
        '200':
          description: Playlists retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'
      security:
        - apiKeyAuth: []
  /playlists/{id}:
    put:
      tags:
        - Playlist
      summary: Edit playlist by id
      description: Update playlist details.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key required to authenticate the request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                desc:
                  type: string
                img:
                  type: string
      responses:
        '200':
          description: Playlist updated successfully
      security:
        - apiKeyAuth: []

    get:
      tags:
        - Playlist
      summary: Get playlist by id
      description: Retrieve a specific playlist by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to authenticate the request.
      responses:
        '200':
          description: Playlist retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Playlist'
      security:
        - apiKeyAuth: []

    delete:
      tags:
        - Playlist
      summary: Delete playlist by id
      description: Remove a playlist.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed for authentication.
      responses:
        '200':
          description: Playlist deleted successfully
      security:
        - apiKeyAuth: []
  /playlists/add-podcast:
    put:
      tags:
        - Playlist
      summary: Add podcast to playlist
      description: Add a podcast to an existing playlist.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key required to authenticate the request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                playlistId:
                  type: string
                podcastId:
                  type: string
      responses:
        '200':
          description: Podcast added to playlist successfully
      security:
        - apiKeyAuth: []
  /playlists/remove-podcast:
    put:
      tags:
        - Playlist
      summary: Remove podcast from playlist
      description: Remove a podcast from a playlist.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key required to authenticate the request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                playlistId:
                  type: string
                podcastId:
                  type: string
      responses:
        '200':
          description: Podcast removed from playlist successfully
      security:
        - apiKeyAuth: []
  /playlists/favourite:
    get:
      tags:
        - Playlist
      summary: Get user's favourite playlists
      description: Retrieve all playlists marked as favourite by the user.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to authenticate the request.
      responses:
        '200':
          description: Favourite playlists retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'
      security:
        - apiKeyAuth: []
  /playlists/random:
    get:
      tags:
        - Playlist
      summary: Get random playlists
      description: Retrieve a random set of playlists.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: API key needed to authenticate the request.
      responses:
        '200':
          description: Random playlists retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'
      security:
        - apiKeyAuth: []
...
components:
  schemas:
    Podcast:
      type: object
      properties:
        name:
          type: string
          example: "Tiktok remix hehe"
        artist:
          type: string
          example: "Duc Dat"
        podcast:
          type: string
          description: "Link to the podcast file"
          example: "podcast.mp3"
        img:
          type: string
          description: "Link to the image file"
          example: "song.jpg"
        duration:
          type: string
          description: "Duration of the podcast in minutes"
          example: "3"
      required:
        - name
        - artist
        - podcast
        - img
        - duration
...
    Playlist:
      type: object
      required:
        - name
        - user
      properties:
        name:
          type: string
          example: "Summer Hits"
        user:
          type: string
          description: "User ID of the playlist owner"
        desc:
          type: string
          example: "A playlist of summer hits"
        podcasts:
          type: array
          items:
            type: string
          description: "Array of podcast IDs included in the playlist"
        img:
          type: string
          description: "URL of the playlist cover image"
...
  securitySchemes:
    auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-auth-token
      description: API key authentication
    adminApiKey:
      type: apiKey
      in: header
      name: x-auth-token
      description: API key authentication for admin endpoints